fix(lexical-editor): address defects
- register editable listener in plugins
- add preview modified
- fix ctrl-a in ImageEditDialog
