+++ packages/cms/lists/views/wysiwyg-editor/style/index.ts
@@ -93,13 +93,15 @@ export const StyleWrapper = styled.div`
   }
 
   .TwreporterTheme__paragraph {
-    ${normalWidthCSS}
-    ${paragraphTextCSS}
-    margin: ${mockup.margin.normal};
-    color: ${colorGrayscale.gray800};
+    &:not(:has(.TwreporterTheme__imageLink)) {
+      ${normalWidthCSS}
+      ${paragraphTextCSS}
+      margin: ${mockup.margin.normal};
+      color: ${colorGrayscale.gray800};
 
-    /* line break */
-    white-space: pre-wrap;
+      /* line break */
+      white-space: pre-wrap;
+    }
   }
   
   .TwreporterTheme__link {
@@ -218,5 +220,4 @@ export const StyleWrapper = styled.div`
   .TwreporterTheme__textCapitalize {
     text-transform: capitalize;
   }
-
 `
