+++ packages/lexical-editor/src/react/plugins/ImagePlugin/Image.scss
@@ -100,7 +100,7 @@
   }
 }
 
-.Image__image_block.small {
+.Image__image_block.right {
   @include desktop-only {
     width: 403px;
   }
